requirements.txt
requests
psutil
python-telegram-bot>=21,<22
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
aioaria2>=1.3.6
python-dotenv>=0.10
tenacity>=6.0.0
python-magic
beautifulsoup4>=4.8.2,<4.8.10
pyrofork
TgCrypto>=1.1.1,<1.1.10
yt-dlp
megasdkrestclient>=0.1.1,<1.0.0
redis>=4.3.1,<5.0.0